sdk/pyproject.toml
[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"

[project]
name = "opendecree"
version = "0.1.0"
description = "Python SDK for OpenDecree — schema-driven configuration management"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Zeev Dreifuss" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "grpcio>=1.68.0,<2",
    "protobuf>=6.30.0,<7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-cov>=6.0",
    "pytest-asyncio>=0.24",
    "mypy>=1.14",
    "mypy-protobuf>=3.6",
    "ruff>=0.8",
    "grpcio-tools>=1.68.0",
]

[project.urls]
Homepage = "https://github.com/zeevdr/decree-python"
Documentation = "https://github.com/zeevdr/decree"
Repository = "https://github.com/zeevdr/decree-python"
Issues = "https://github.com/zeevdr/decree-python/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100
exclude = ["src/opendecree/_generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "RUF"]

[tool.ruff.lint.per-file-ignores]
"src/opendecree/_generated/**" = ["ALL"]
"src/opendecree/__init__.py" = ["E402"]
"tests/**" = ["S101"]

[tool.mypy]
python_version = "3.11"
strict = true

[[tool.mypy.overrides]]
module = "opendecree._generated.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "grpc.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["opendecree"]
omit = ["src/opendecree/_generated/*"]

[tool.coverage.report]
fail_under = 80
